# backend/openapi.yaml
openapi: 3.0.3
info:
  title: BANKERCHANGER API
  version: 0.2.0
  description: REST API for the BANKERCHANGER decentralized boxing prediction market.
servers:
  - url: http://localhost:3001
    description: Local development
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    oracleKey:
      type: apiKey
      in: header
      name: X-Oracle-Key
  schemas:
    Error:
      type: object
      properties:
        success: { type: boolean, example: false }
        message: { type: string }
    ErrorResponse400:
      type: object
      properties:
        success: { type: boolean, example: false }
        message: { type: string, description: "Bad request. Invalid query parameters, request body, or format." }
        code: { type: string, example: "INVALID_REQUEST" }
      required: [success, message]
...
    ErrorResponse429:
      type: object
      properties:
        success: { type: boolean, example: false }
        message: { type: string, description: "Rate limit exceeded. Check Retry-After header for reset time." }
        code: { type: string, example: "RATE_LIMIT_EXCEEDED" }
        retryAfter: { type: integer, description: "Seconds until the rate limit window resets." }
      required: [success, message]
    ErrorResponse500:
      type: object
      properties:
        success: { type: boolean, example: false }
        message: { type: string, description: "Internal server error. Please try again later." }
        code: { type: string, example: "INTERNAL_SERVER_ERROR" }
        requestId: { type: string, description: "Unique identifier for this error instance for debugging." }
      required: [success, message]
...
    Pagination:
      type: object
      properties:
        page:       { type: integer }
        limit:      { type: integer }
        total:      { type: integer }
        totalPages: { type: integer }
        hasNext:    { type: boolean }
        hasPrev:    { type: boolean }
    GovernanceProposal:
      type: object
      properties:
        id:           { type: string }
        type:         { type: string }
        value:        { type: string }
        description:  { type: string }
        status:
          type: string
          enum: [Active, Passed, Failed, Executed]
        proposer:     { type: string }
        votesFor:     { type: number }
        votesAgainst: { type: number }
        votesAbstain: { type: number }
        createdAt:    { type: string, format: date-time }
        expiresAt:    { type: string, format: date-time }
...
paths:
...
  /api/governance/proposals:
    get:
      summary: List all governance proposals (paginated)
      operationId: listProposals
      tags: [Governance]
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1, minimum: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 20, maximum: 100, minimum: 1 }
        - name: status
          in: query
          schema:
            type: string
            enum: [active, passed, failed, executed]
      responses:
        '200':
          description: Paginated list of proposals
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/GovernanceProposal' }
                  pagination: { $ref: '#/components/schemas/Pagination' }
        '400':
          description: Invalid query parameters
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse400' }
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse429' }
        '500':
          description: Internal server error
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse500' }
  /api/governance/proposals/{proposal_id}:
    get:
      summary: Get a single governance proposal by ID
      operationId: getProposal
      tags: [Governance]
      parameters:
        - name: proposal_id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Proposal details
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:    { $ref: '#/components/schemas/GovernanceProposal' }
        '404':
          description: Proposal not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse429' }
        '500':
          description: Internal server error
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse500' }
...